Refine the directory structure and sketch some of the classes.
